do not sleep in packet strobe in packet_to_stream example
Try to make the example work with a single-threaded scheduler by not
sleeping in the packet strobe block.

Also add requestStop() any time we emit a fatal error message, since
returning gr::work::Status::ERROR doesn't seem to be sufficient to
terminate the flowgraph and get the error message in the runAndWait()
return value.

Signed-off-by: Daniel Estévez <daniel@destevez.net>