fix unit tests after gnuradio4 update
Now the flowgraphs which have a block with an Async port,
such as PduToTaggedStream, won't terminate on their own even if the
input to this block terminates (because in principle such block could
still produce some output "for free" with no input). Therefore, the
tests that use these blocks need to be stopped manually with a
scheduler message.

Additionally, now blocks that want to overwrite tags need to use
TPP_CUSTOM. It is not possible to let the scheduler propagate tags
automatically and publish a tag with to overwrite the map entries.

Finally, a timeout of 30 seconds has been added for all the tests.

One of the qa_pdu tests is failing, but this seems to be a tag
propagation bug in the runtime.

Signed-off-by: Daniel Estévez <daniel@destevez.net>