CrcAppend: do not consume all input if there's no room for CRC
Previously, the CrcAppend block consumed all its input and stored
internally the CRC to be produced in a later call if there's no
room in the output buffer in the current call. This gives problems,
because the runtime doesn't understand that the block can produce
output "for free", without needing any input, in this state.

To get around this, the last item in the input is not consumed
until there is room in the output to produce the CRC fully.

Signed-off-by: Daniel Estévez <daniel@destevez.net>