refactor SymbolFilter to account for clock jumps
This refactors the SymbolFilter block to handle correctly the case
of back-to-back packets. In this case, the clock can jump by one
sample at the beginning of one packet (specially due to the fine
time estimate). This refactor ensures that there are no repeated
or missing output symbols due to this jump and that the tags end
up in the correct location.

Signed-off-by: Daniel Estévez <daniel@destevez.net>