add delay to coarse frequency correction
This fixes a potential decoding failure with back-to-back packets.
If the frequency detected using the syncword is slightly wrong and
the packet is long, the packet will accumulate significant phase
error in the coarse frequency correction block. At the end of the
packet, if the phase is jumped back to zero at the start of the
syncword (which "bleeds into" the last symbols of the packet due
to the RRC filter), then the phase jump will cause bit errors. The
solution is to delay the application of the syncword tags by the
coarse frequency correction block.

With this change, the qa_loopback is now working in all the
configurations.

Signed-off-by: Daniel Estévez <daniel@destevez.net>