add latency management blocks
This adds simple latency management blocks, but the solution doesn't
work due to the buffers between the TUN Source and the Packet Limiter
and between the Packet Counter and the Packet To Stream. It is necessary
to integrate the latency management functionality on the TUN Source
and the Packet To Stream, since the buffer sizes cannot be made very
small at the moment.

Signed-off-by: Daniel Estévez <daniel@destevez.net>