syncword_detection: increase threshold slightly
Running "packet_transceiver 60 0.000 0.1" and a long iperf3 TCP
transfer, I have noticed that some particular TCP packets would
not be decoded correctly (ever or after many retransmissions). This
would cause the TCP transfer to stall. The problem is that a syncword
is detecting mid-packet due to the pattern of the data in the packet.
This causes bit errors in the decoding, because of the syncword
wipeoff, and potentially other problems (frequency errors, etc.).

Increasing the syncword threshold slightly seems to fix this. Now
a 300 second TCP transfer goes through without problems.

Signed-off-by: Daniel Estévez <daniel@destevez.net>