pdu_to_tagged_stream: add enable_count setting
There was a bug in PduToTaggedStream. The count output is optional,
but the block puts items into this output even if it is not connected.
This causes the output buffer to fill up and then the block essentially
gets stuck, because it doesn't have room for more output. It seems
that this problem was masked by a bug in the PublishableOutputRange
that sometimes reset the state of the buffer to zero.

To fix this bug, an enable_count setting is added. It defaults to false,
which means that PduToTaggedStream will not generate a count output. In
fact, the count output is not used at the moment in the flowgraphs in
apps/ or examples/

Signed-off-by: Daniel Estévez <daniel@destevez.net>